+++ net/server.go
@@ -2,10 +2,7 @@ package net
 
 import (
 	"blockchain_go/blockchain"
-	"blockchain_go/transaction"
-	"bytes"
 	"fmt"
-	"io"
 	"io/ioutil"
 	"log"
 	"net"
@@ -19,118 +16,6 @@ var nodeAddress string
 var miningAddress string
 var CenterNode string = "localhost:3000"
 var knownNodes = []string{CenterNode}
-
-func commandToBytes(command string) []byte {
-	var bytes [commandLength]byte
-
-	for i, c := range command {
-		bytes[i] = byte(c)
-	}
-
-	return bytes[:]
-}
-
-func bytesToCommand(bytes []byte) string {
-	var command []byte
-
-	for _, b := range bytes {
-		if b != 0x0 {
-			command = append(command, b)
-		}
-	}
-
-	return fmt.Sprintf("%s", command)
-}
-
-func extractCommand(request []byte) []byte {
-	return request[:commandLength]
-}
-
-func requestBlocks() {
-	for _, node := range knownNodes {
-		sendGetBlocks(node)
-	}
-}
-
-func sendAddr(address string) {
-	nodes := addr{knownNodes}
-	nodes.AddrList = append(nodes.AddrList)
-	payload := gobEncode(nodes)
-	request := append(commandToBytes("addr"), payload...)
-
-	sendData(address, request)
-}
-
-func sendBlock(addr string, b *blockchain.Block) {
-	data := block{b.Serialize()}
-	payload := gobEncode(data)
-	request := append(commandToBytes("block"), payload...)
-
-	sendData(addr, request)
-}
-
-func sendData(addr string, data []byte) {
-	conn, err := net.Dial(protocol, addr)
-	if err != nil {
-		fmt.Printf("%s is not available\n", addr)
-		var updatedNodes []string
-
-		for _, node := range knownNodes {
-			if node != addr {
-				updatedNodes = append(updatedNodes, node)
-			}
-		}
-
-		knownNodes = updatedNodes
-
-		return
-	}
-	defer conn.Close()
-
-	_, err = io.Copy(conn, bytes.NewReader(data))
-	if err != nil {
-		log.Panic(err)
-	}
-}
-
-func sendInv(address, kind string, items [][]byte) {
-	inventory := inv{kind, items}
-	payload := gobEncode(inventory)
-	request := append(commandToBytes("inv"), payload...)
-
-	sendData(address, request)
-}
-
-func sendGetBlocks(address string) {
-	payload := gobEncode(getblocks{})
-	request := append(commandToBytes("getblocks"), payload...)
-
-	sendData(address, request)
-}
-
-func sendGetData(address, kind string, id []byte) {
-	payload := gobEncode(getdata{kind, id})
-	request := append(commandToBytes("getdata"), payload...)
-
-	sendData(address, request)
-}
-
-func SendTx(addr string, tnx *transaction.Transaction) {
-	data := tx{tnx.Serialize()}
-	payload := gobEncode(data)
-	request := append(commandToBytes("tx"), payload...)
-
-	sendData(addr, request)
-}
-
-func sendVersion(addr string, bc *blockchain.Blockchain) {
-	bestHeight := bc.GetBestHeight()
-	payload := gobEncode(version{nodeVersion, bestHeight})
-
-	request := append(commandToBytes("version"), payload...)
-
-	sendData(addr, request)
-}
 
 func handleConnection(conn net.Conn, bc *blockchain.Blockchain) {
 	defer conn.Close()
